UML diagrams now include as PlantUML sources directly (not as png-files) .
This solve a memory issue during creating a 'PDF'.

Issue during creation a  'PDF'::
Create a PDF of a big AsciiDoc file (adoc-main-file),
which consists of several .adoc files (adoc-sub-files)
and include UML diagrams as `png`-files through the `image` commands.
The issue: The PDF of the adoc-main-file may not contain all diagrams,
but the PDFs created from the adoo-sub-files will.